requirements.txt
Flask
psycopg2-binary
flask-cors
flask-jwt-extended
